requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.38
pymysql==1.1.0
APScheduler==3.11.0
Flask-APScheduler==1.13.1
gunicorn==23.0.0
cryptography
celery
redis
google-generativeai